+++ attribute/key_test.go
@@ -108,6 +108,11 @@ func TestEmit(t *testing.T) {
 			v:    attribute.StringSliceValue([]string{"foo", "bar"}),
 			want: `["foo","bar"]`,
 		},
+		{
+			name: `test Key.Emit() can emit a string representing self.SLICE`,
+			v:    attribute.SliceValue([]attribute.Value{attribute.BoolValue(true), attribute.IntValue(42)}),
+			want: `[{1 1  <nil>} {2 42  <nil>}]`,
+		},
 	} {
 		t.Run(testcase.name, func(t *testing.T) {
 			// proto: func (v attribute.Value) Emit() string {